some refactoring were added due to some improvements recomended in the last PR review
